+++ wordless/wl_ngram_generator.py
@@ -921,6 +921,9 @@ class Wl_Worker_Ngram_Generator(wl_threading.Wl_Worker):
 
             # Total
             if len(files) > 1:
+                if not self._running:
+                    raise wl_excs.Wl_Exc_Aborted(self.main)
+
                 texts.append(wl_texts.Wl_Text_Total(texts))
 
                 self.ngrams_freq_files.append(sum([