+++ CHANGELOG.md
@@ -23,6 +23,7 @@
 - File Area: Add support for .srt files
 - Settings: Add Settings - Tables - Miscellaneous Settings
 - Utils: Add Wordless's character tokenizer
+- Work Area: Add Profiler - Export all tables
 - Work Area: Add Work Area - Table Settings - Show total
 
 ### ✨ Improvements